Move other files.
