+++ crates/core/cache/src/items/run_target_state.rs
@@ -3,7 +3,7 @@ use crate::helpers::{is_readable, is_writable};
 use moon_archive::{untar, TarArchiver};
 use moon_error::MoonError;
 use moon_logger::{color, trace};
-use moon_utils::{fs, time};
+use moon_utils::{fs, json, time};
 use serde::{Deserialize, Serialize};
 use std::path::{Path, PathBuf};
