+++ crates/core/cache/src/items/tool_state.rs
@@ -2,7 +2,7 @@ use crate::cache_item;
 use crate::helpers::{is_readable, is_writable};
 use moon_error::MoonError;
 use moon_logger::{color, trace};
-use moon_utils::{fs, time};
+use moon_utils::{fs, json, time};
 use serde::{Deserialize, Serialize};
 use std::path::{Path, PathBuf};
